+++ testdata/codegen_matrix/fixtures/internal_fields/src/ent/assignment.ts
@@ -1,0 +1,8 @@
+import { AlwaysAllowPrivacyPolicy } from "@snowtop/ent";
+import { AssignmentBase } from "./internal";
+
+export class Assignment extends AssignmentBase {
+  getPrivacyPolicy() {
+    return AlwaysAllowPrivacyPolicy;
+  }
+}