+++ testdata/codegen_matrix/fixtures/internal_fields/src/ent/assignment/actions/delete_assignment_action.ts
@@ -1,0 +1,8 @@
+import { DeleteAssignmentActionBase } from "../../generated/assignment/actions/delete_assignment_action_base";
+import { deriveAssignment } from "./derive_assignment";
+
+export default class DeleteAssignmentAction extends DeleteAssignmentActionBase {
+  getTriggers() {
+    return [{ changeset: deriveAssignment }];
+  }
+}